Main Html file
This is the main HTML file for my advanced portfolio website. It serves as the entry point for the site, integrating HTML, CSS, Bootstrap, and JavaScript to create a responsive and visually appealing user experience. The file includes the navigation bar, main content sections, and footer. It's designed to showcase my skills, projects, and contact information effectively.